COMMERCE-10759 Refactor to a displayContext
